Fix race condition in multi-channel ingestion
When the first alert for a given diaObject was accepted by only one
channel, datapoints in that alert (including the diaObject entry)
would not be added to other channels when present in subsequent
alerts.  This happened _only_ if the datapoints were found in the
t0 collection.  Prevent this by always emitting dp content from the
database, and projecting alert dps the same way.