LSSTMongoMuxer: filter t0 docs by ttl
Datapoints may be deleted by ttl cleanup between the time the find
query is issued and the updates from ingestion are committed to the
database.  Insulate against this by only returning datapoints that
will not be cleaned up for a configurable time.  This causes
datapoints in the current alert that are already in the database
but about to be deleted to be upserted again, with an updated ttl.